requirements.txt
Django==5.0.7
streamlit==1.36.0
beautifulsoup4==4.12.3
requests==2.32.3
Pillow==10.4.0
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
firebase-admin==6.5.0
google-generativeai==0.7.1

